+++ R/make_unique_id.R
@@ -1,15 +1,26 @@
 #' Create a column with unique id per row values
 #'
+#' @description
+#' `r lifecycle::badge("deprecated")`
+#'
+#' The underlying Polars function isn't guaranteed to give the same results
+#' across different versions. Therefore, this function will be removed and has
+#' no replacement in `tidypolars`.
+#'
 #' @inheritParams fill.polars_data_frame
 #' @param new_col Name of the new column
 #'
 #' @export
-#' @examplesIf require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)
-#' mtcars |>
-#'   as_polars_df() |>
-#'   make_unique_id(am, gear)
-
 make_unique_id <- function(.data, ..., new_col = "hash") {
+  lifecycle::deprecate_warn(
+    when = "0.16.0",
+    what = "make_unique_id()",
+    details = c(
+      "This has no guarantee of giving the same results across Polars versions.",
+      "It has no replacement in `tidypolars`."
+    ),
+    always = TRUE,
+  )
   check_polars_data(.data)
   if (new_col %in% names(.data)) {
     cli_abort(